Rename EditPopUp
